+++ pyproject.toml
@@ -70,12 +70,11 @@ all = [
 [dependency-groups]
 dev = [
     "black>=24.8.0; python_version >= '3.10'",
-    "isort==7.0.0; python_version >= '3.10'",
+    "isort>=7.0.0; python_version >= '3.10'",
     "flake8>=5.0.4",
     "flake8-pyproject>=1.2.3",
     "mypy>=1.14.1",
     "sphinx>=7.1.2",
-    "isort>=5.13.2",
 ]
 
 [tool.scikit-build.metadata.version]
