+++ src/scaler/scheduler/controllers/policies/simple_policy/scaling/capability_scaling.py
@@ -9,9 +9,9 @@ from scaler.protocol.capnp import (
     WorkerManagerCommandType,
     WorkerManagerHeartbeat,
 )
-from scaler.protocol.helpers import build_scaling_manager_status
 from scaler.scheduler.controllers.policies.simple_policy.scaling.mixins import ScalingPolicy
 from scaler.scheduler.controllers.policies.simple_policy.scaling.types import WorkerManagerSnapshot
+from scaler.scheduler.controllers.worker_manager_utilties import build_scaling_manager_status
 from scaler.utility.identifiers import WorkerID
 from scaler.utility.snapshot import InformationSnapshot
