+++ src/scaler/scheduler/controllers/worker_manager_controller.py
@@ -13,10 +13,11 @@ from scaler.protocol.capnp import (
     WorkerManagerHeartbeat,
     WorkerManagerHeartbeatEcho,
 )
-from scaler.protocol.helpers import build_scaling_manager_status, capabilities_to_dict
+from scaler.protocol.helpers import capabilities_to_dict
 from scaler.scheduler.controllers.config_controller import VanillaConfigController
 from scaler.scheduler.controllers.mixins import PolicyController, TaskController, WorkerController
 from scaler.scheduler.controllers.policies.simple_policy.scaling.types import WorkerManagerSnapshot
+from scaler.scheduler.controllers.worker_manager_utilties import build_scaling_manager_status
 from scaler.utility.identifiers import WorkerID
 from scaler.utility.mixins import Looper, Reporter
 from scaler.utility.snapshot import InformationSnapshot
